refactor(dashboard): split the chart/window hub out of views.py (#1105 Phase 3, develop-v2)
D1 of the dashboard cluster: the window/range hub every other view cluster leans on
moves to web/charts.py, and its tests move with it.

Moved (byte-identical, three contiguous blocks): the range/point/tension tiers and
parse_window/canonical_window, build_chart and the range filters, the window-duration
and downsampling helpers, and the four marker series (shares, events, raffle, payouts).

views.py stays the facade — server.py's `parse_window`/`canonical_window` and
worker_detail.py's `_filter_events` imports are untouched and still resolve through it.

Tests move 1:1: TestChart, TestParseWindow, TestChartEvents, TestChartRaffle and
TestChartPayouts to tests/web/test_charts.py, same 2120 collected IDs with only the
module path changed.
